edit shell_sort.c

+++ 100-shell_sort.c
@@ -8,7 +8,7 @@
   */
 void shell_sort(int *array, size_t size)
 {
-	unsigned int i = 0, j = 0, gap = 0;
+	size_t i = 0, j = 0, gap = 0;
 	int aux = 0;
 
 	if (array == NULL || size < 2)
@@ -35,3 +35,4 @@ void shell_sort(int *array, size_t size)
 		print_array(array, size);
 	}
 }
+
